pyproject.toml
[project]
name = "ai-env"
version = "0.1.0"
description = "AI 개발 환경 및 토큰 통합 관리 도구"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Namjung Won", email = "megan.won@kakaocorp.com" }
]
keywords = ["ai", "cli", "mcp", "claude", "gemini", "devtools"]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    # Skill dependencies
    "jira>=3.6.0",
    "html2text>=2024.2.26",
    "mistune>=3.0",
    "trino>=0.328.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.3",
    # Test dependency (so `uv run pytest` uses project env, not system pytest)
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
ai-env = "ai_env.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_env"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PT"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "click.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["ai_env.cli", "ai_env.cli.*"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["tests.*", "scripts.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "--cov=ai_env --cov-report=term-missing --cov-config=pyproject.toml"

[tool.coverage.run]
source = ["src/ai_env"]
omit = [
    "src/ai_env/cli/__init__.py",
]

[tool.coverage.report]
fail_under = 65
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
